Fling squash pt2
Navigation: All the bars have a Pulse! [1/4]
Pulse is now managed as a persistent member of
NavigationController and the instance is passed to
bars.

Pulse: implementation of softkey pulse start/stop animation
A low alpha of 0.3f should do the trick to make keys visible
while not being intrusive to the Pulse rendering. Try to make
Pulse fade play nice with lights out.

Fling: BatteryBar support

SystemUI: Fix navbar drawing issues in landscape
Originally StatusBarWindowView was responsible for laying out the scrims
and making sure they are full screen (behind navbar), while the other
views were respecting the navbar insets.

This was broken for two reasons for us:

 1) our scrims are more complex and aren't always direct subchildren of
    the StatusBarWindowView (mainly to display the visualizer correctly)

 2) we declared fitSystemWindows=false, so the logic to add insets was
    never actually executing.

We now split out the logic for insets into its own class and apply it to
the scrims properly.

Ref: CYNGNOS-1502
Change-Id: I5b8295d7a64b72dc911fce6d529936592f24e98b
Signed-off-by: Roman Birg <roman@cyngn.com>
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>
creds romanbb

statusbar: add screenPinningStateChanged(boolean enabled) api and pip…
…e to navigation bar

Up until now there was no api to know exactly when tasks became locked from screen
pinning. When ActivityStackSupervisor gets the call from ActivityManager, send it
down the statusbar pipes as well.

Smartbar: Cripple screen pinning confirmation and get right to it
Due to our new task lock hint api an instructional dialog is
no longer needed or desired. Also, it is rather impractical to
keep the dialog but create some kind of "on the fly" dynamic
layout in the dialog. So when the screen pinning command comes
down the pipes, just fire it.